+++ R/extract.R
@@ -14,7 +14,8 @@
 #'
 #' - `extract_parameter_set_dials()` returns a set of dials parameter objects.
 #'
-#' @param x A parsnip `model_fit` object.
+#' @param x A parsnip `model_fit` object or a parsnip `model_spec` object.
+#' @param id A single string for the `id` of the parameter.
 #' @param ... Not currently used.
 #' @details
 #' Extracting the underlying engine fit can be helpful for describing the
@@ -111,4 +112,10 @@ eval_call_info <-  function(x) {
     res <- NA
   }
   res
+}
+
+#' @export
+#' @rdname extract-parsnip
+extract_parameter_dials.model_spec <- function(x, id, ...) {
+  extract_parameter_dials(extract_parameter_set_dials(x), id)
 }