+++ R/reexports.R
@@ -38,3 +38,7 @@ hardhat::extract_fit_engine
 #' @importFrom hardhat extract_parameter_set_dials
 #' @export
 hardhat::extract_parameter_set_dials
+
+#' @importFrom hardhat extract_parameter_dials
+#' @export
+hardhat::extract_parameter_dials